refactor(pages): remove components, which are not related to routes from next.js pages folder
